requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
dill==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.5
idna==3.10
isort==5.13.2
joblib==1.4.2
mccabe==0.7.0
mysql-connector-python==9.1.0
numpy==2.0.2
pandas==2.2.3
platformdirs==4.3.6
pydantic==2.10.1
pydantic_core==2.27.1
pylint==3.3.1
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
threadpoolctl==3.5.0
tomli==2.1.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
